Factor out common RingBuffer functionality into RingBufferBase
